- PR comments - Note: When we passing the FirebaseUser object over to the conversation_page using provider a getter is called on null object. Let test function for recreating both scenarios. (TODO comment.)
