+++ lib/components/send_message_button.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class SendButton extends StatelessWidget {
+  final String text;
+  final VoidCallback callback;
+
+  const SendButton({Key key, this.text, this.callback}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return FlatButton(
+      onPressed: callback,
+      child: Text(text),
+      color: Colors.amber,
+    );
+  }
+}
